test-requirements.txt
pytenable>=0.3.10
arrow>=0.13.0
google-auth>=1.6.3
Click>=7.0
pytest>=3.6
pytest-cov
pytest-vcr
pytest-datafiles